@@ -1,10 +1,18 @@
 <script>
   import Dispatcher from '../../dispatcher.js'
 
-  export let regionNames
+  let { regionNames } = $props()
 
   let ctrlKey = false
   let metaKey = false
+
+  let inputValue = $state('')
+
+  // this is because the html datalist does not support a fuzzy search with macrons
+  // so we just have to remove them from the search ui
+  const removeSpecialCharacters = (str) => {
+    return str.normalize('NFD').replace(/[\u0300-\u036f]/g, '')
+  }
 
   const onKeyPress = (e) => {
     if (e.ctrlKey !== undefined) ctrlKey = e.ctrlKey
@@ -15,10 +23,12 @@
   const onSearch = async (e) => {
     const data = await regionNames
     if (e.currentTarget === null) return
-    const { value } = e.currentTarget
+    const value = inputValue
 
     // checks to make sure they used a precanned one
-    const match = data.find((region) => region.name === value)
+    const match = data.find(
+      (region) => removeSpecialCharacters(region.name) === value
+    )
     if (match) {
       if (match.id.startsWith('TZ')) {
         Dispatcher.dataSegment = Dispatcher.dataSegment.replace(
@@ -48,24 +58,25 @@
   <input placeholder="Loading..." />
 {:then regions}
   <input
+    bind:value={inputValue}
     list="search-choice"
     placeholder="Search areas..."
-    on:select={onSearch}
-    on:change={onSearch}
-    on:keydown={onKeyPress}
-    on:keyup={onKeyPress}
-    on:focus={onFocus}
+    onselect={onSearch}
+    onchange={onSearch}
+    onkeydown={onKeyPress}
+    onkeyup={onKeyPress}
+    onfocus={onFocus}
   />
   <datalist id="search-choice">
     {#each regions as region}
-      <option value={region.name} />{/each}
+      <option value={removeSpecialCharacters(region.name)}></option>{/each}
   </datalist>
 {/await}
 
 <style>
   input {
-    background: #222;
-    color: #fff;
+    background: var(--surface-bg);
+    color: var(--surface-text);
     font-family: 'Fira Sans';
     font-size: 15px;
     padding: 0.35rem 0.5rem;
@@ -74,9 +85,10 @@
     margin-bottom: 3px;
     outline: 0;
     border-radius: 3px;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.15);
   }
   input:focus {
-    background: #111;
-    border-color: #000;
+    background: var(--surface-bg);
+    border-color: var(--surface-text-interactive);
   }
 </style>
