@@ -1,4 +1,6 @@
 <script>
+  import { run } from 'svelte/legacy'
+
   import expand from '/static/expand.svg'
   import ason from '/static/css/ason.png'
   import { transformFilename } from '../../data.js'
@@ -6,13 +8,12 @@
   import { getSource } from '../../sources.js'
   import ModeToggle from './ModeToggle.svelte'
 
-  export let title, firstRegion, populationLabel, populationLink
+  let { title, firstRegion, populationLabel, populationLink, populationCount } =
+    $props()
 
   const source = getSource()
 
-  $: path = transformFilename(firstRegion)
-
-  let isControlsHidden = false
+  let isControlsHidden = $state(false)
 
   const triggerClose = () => {
     Dispatcher.setRegions([])
@@ -33,28 +34,41 @@
     }
   }
   const triggerSecondarySegment = (segment) => () => {
-    // more ason specific stuff
     if (Object.keys(Dispatcher.concordance).length > 0) {
-      Dispatcher.currentRegion = Dispatcher.currentRegion
-        .map(
-          (i) =>
-            Dispatcher.concordance[i][`${segment}-sa2`] ||
-            Dispatcher.concordance[i][`${segment}-dzn`]
+      // more ason specific stuff
+      if (source.brandingClass === 'ason') {
+        Dispatcher.currentRegion = Array.from(
+          new Set(
+            Dispatcher.currentRegion
+              .map(
+                (i) =>
+                  Dispatcher.concordance[i][`${segment}-sa2`] ||
+                  Dispatcher.concordance[i][`${segment}-dzn`]
+              )
+              .flat()
+          )
         )
-        .flat()
+      } else {
+        Dispatcher.currentRegion = Array.from(
+          new Set(
+            Dispatcher.currentRegion
+              .map((i) => Dispatcher.concordance[i][`${segment}-sa2`])
+              .flat()
+          )
+        )
+      }
     }
     const newSegment = segment.toLowerCase()
     const primary = Dispatcher.dataSegment.split('-')[1]
     setSegmentWithMode([newSegment, primary].join('-'), selection)
   }
-  let currentSegment = source.segments[0]
+  let currentSegment = $state(source.segments[0])
   const loadBlocks = () => {
     currentSegment = Dispatcher.dataSegment
   }
   Dispatcher.bind('load-blocks', loadBlocks)
 
-  let selection = []
-  $: setSegmentWithMode(Dispatcher.dataSegment, selection)
+  let selection = $state([])
   const setSegmentWithMode = (segment, selectedModes) => {
     let finalSegment = segment.split('-').slice(0, 2).join('-')
     if (selectedModes.length > 0) {
@@ -62,6 +76,15 @@
     }
     Dispatcher.setSegment(finalSegment)
   }
+  // removes the macrons for the stats nz link
+  const path = $derived(
+    transformFilename(firstRegion)
+      .normalize('NFD')
+      .replace(/[\u0300-\u036f]/g, '')
+  )
+  $effect(() => {
+    setSegmentWithMode(Dispatcher.dataSegment, selection)
+  })
 </script>
 
 <div class="nav-header" class:ason={source.brandingClass === 'ason'}>
@@ -76,12 +99,12 @@
       <button
         title="Learn More"
         class="btn-expand"
-        on:click={() =>
+        onclick={() =>
           document.getElementById('app').classList.toggle('map-view')}
       >
         <img alt="Toggle Details" src={expand} />
       </button>
-      <button title="Close Location" class="btn-close" on:click={triggerClose}>
+      <button title="Close Location" class="btn-close" onclick={triggerClose}>
         <svg
           xmlns="http://www.w3.org/2000/svg"
           height="24"
@@ -108,7 +131,7 @@
                 <a
                   href="#"
                   class="btn-segment"
-                  on:click={triggerSecondarySegment(control)}
+                  onclick={triggerSecondarySegment(control)}
                   class:selected={currentSegment.split('-')[0] ===
                     control.toLowerCase()}>{control}</a
                 >
@@ -121,16 +144,16 @@
         {#if populationLink}
           <a
             class="population-link"
-            href="https://www.stats.govt.nz/tools/2018-census-place-summaries/{path}#population-and-dwellings`"
+            href="https://tools.summaries.stats.govt.nz/places/SA2/{path}"
             target="_blank"
             rel="noopener noreferrer"
           >
             <strong class="population-label">{populationLabel}</strong>
-            <span class="population-count" />
+            <span class="population-count">{populationCount}</span>
           </a>
         {:else}
           <strong class="population-label">{populationLabel}</strong>
-          <span class="population-count" />
+          <span class="population-count">{populationCount}</span>
         {/if}
       </p>
     </div>
@@ -141,7 +164,7 @@
             <a
               href="#"
               class="btn-segment"
-              on:click={triggerSegment(control)}
+              onclick={triggerSegment(control)}
               class:selected={(source.detailsSecondaryControls != null
                 ? currentSegment.split('-')[1]
                 : currentSegment
@@ -160,6 +183,115 @@
 </div>
 
 <style>
+  .nav-header {
+    padding: 1rem var(--sidebar-padding) 1rem;
+    background: var(--surface-bg-subtle);
+    border-bottom: var(--border);
+    position: sticky;
+    top: 0;
+    margin-bottom: 2rem;
+    box-sizing: border-box;
+    z-index: 2;
+  }
+
+  .nav-header-flex {
+    display: flex;
+  }
+
+  .nav-header .title {
+    flex: 1;
+    overflow: hidden;
+    padding-right: 1rem;
+  }
+
+  .nav-header h2 {
+    margin: 0.15rem 0 0.25rem;
+    font-size: 2rem;
+    line-height: 2.4rem;
+    white-space: pre;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    text-transform: capitalize;
+  }
+
+  .nav-header a {
+    color: var(--surface-text);
+    text-decoration: none;
+  }
+  .nav-header a:hover {
+    color: var(--surface-text-subtle);
+    text-decoration: underline;
+  }
+
+  .nav-header p {
+    font-size: 0.95rem;
+    margin: 0;
+  }
+
+  .nav-header .controls {
+    text-align: right;
+  }
+
+  .nav-header button {
+    border: 0;
+    background: transparent;
+    color: #ddd;
+    font-family: inherit;
+    appearance: none;
+    -moz-appearance: none;
+    -webkit-appearance: none;
+  }
+
+  .nav-header button {
+    text-align: center;
+    padding: 0.25rem;
+    margin-bottom: 0.55rem;
+    background: #111;
+    border-radius: 50%;
+    transition: 100ms ease all;
+  }
+
+  .nav-header button:hover {
+    background: #444;
+  }
+
+  .nav-header button:active {
+    background: #333;
+  }
+
+  .nav-header button svg,
+  .nav-header button img {
+    vertical-align: top;
+    width: 24px;
+    height: 24px;
+  }
+
+  .nav-header .controls a {
+    display: inline-block;
+    font-size: 0.9rem;
+    padding: 0 0.25rem;
+  }
+
+  .nav-header a.selected {
+    color: var(--surface-text);
+    font-weight: bold;
+  }
+
+  .nav-header button.btn-expand {
+    display: none;
+  }
+
+  @media (max-width: 1020px) {
+    .nav-header button.btn-expand {
+      transform: translate(36px, 0);
+      display: inline-block;
+    }
+    .nav-header button.btn-close {
+      opacity: 0;
+      pointer-events: none;
+    }
+  }
+
   nav ul {
     display: inline-block;
     list-style-type: none;
@@ -176,7 +308,7 @@
   .primary-controls li:not(:last-child)::after,
   .secondary-controls li:not(:last-child)::after {
     content: '·';
-    margin-right: 3px;
+    margin: 0 3px;
   }
   .ason img.logo {
     display: none;
